Improve marc rules retrieval and add more rules
Initial marc rules retrieval was done based on the order of the rules in the file which was not going to be sustainavble longerm. This provides a `getRule` method that allows retrieving by MARC tag string which is at least an improvement even if does not end up as our final form.

Adds handling for content_type which demonstrates our ability to pull from the leader (and then uses a lookup table to translate the byte to a text representaton.

Adds handling for the year of publication which demonstrates our ability to take a portion of a returned subfield. :see_no_evil:

Documents that our implementation of 856 is naive and not what we need to do longterm.