Refactor to updated rules file format and add some additional fields
This updates to a new, more programmatically actionable and definitive
version of marc_rules.json.

It also adds documentation to some methods and tries to use better
names throughout.

To test some aspects of the new marc_rules format, it was necessary to
not just refactor the existing code, but also add some additional field
parsing that we wanted to be able to do with the new format.