Simplify pipeline logic by removing waitgroup
This removes the need for sync.WaitGroup. A parser gets a record channel
that it closes when it's done parsing. The consumer reads from the same
channel and writes to a done channel when it has finished consuming. The
Process function sets things up. Ultimately, most of the logic in the
Process function should get split off into a non-marc specific pipeline
package.