export valid json to stdout
Log writes to stderr, print logs to stdout.

I wanted to be able to export our json to a file and this allows that. It also tweaks the json creation slightly to create valid json by creating the array syntax and adding comma separators.

I also slipped the original consumer back in because I felt for the stage I am at in development I really just want a quick dump of information and not full data structures or es writes. I can be swayed this is not useful later... but I kinda do want this in for now if it doesn't make anyone super angry.