Additional MARC mappings
Completes https://mitlibraries.atlassian.net/browse/DIP-139 with some
exceptions that will be opened on individual tickets for resolving in
a future sprint.

Adds initial Consumer testing by providing an io.Writer to the Consumer functions that we default to stdout in normal use but can override to a buffer in testing. ES consumer will require a different approach.