Separate out pipeline functionality.
This commit primarily adds a Pipeline struct that handles construction
and execution of the Pipeline, along with three interfaces. A Pipeline
consists of a single Generator, zero or more Transformers and a single
Consumer.

The main goal was to decouple the pipeline functionality from the
parsers, which is where it was living. The pipeline is now entirely
created and run by main, and none of the parsers or consumers even know
about the pipeline.

The other big change is to the pipeline itself handle hooking up
channels between stages. It's no longer necessary to create any channels
up front. The pipeline's Run() method will return an empty channel that
can be read from to determine when the pipeline has completed.

This decoupling makes the test setup considerably easier. Slotting in
new Generators, Transformers and Consumers should only require minimal
changes to main going forward.