Add AWS v4 signing
This adds support for v4 signing of the Elasticsearch requests, which is
required for our Elasticsearch instance. A new command line flag has
been added (--v4) to enable signing. The order in which it will look for
signing credentials is:

1. In environment variables (AWS_ACCESS_KEY_ID or AWS_ACCESS_KEY for the
key ID and AWS_SECRET_ACCESS_KEY or AWS_SECRET_KEY for the secret key).
2. From a credentials file in $HOME/.aws/credentials.
3. From an EC2 role provider.

I'm not sure how/if this last one will work. It will probably need some
tweaking once we start running on Fargate.

To avoid having to copy/paste a bunch, I've moved the url, index and v4
command line parameters to be global params.

The current release of olivere/elastic has a bug in how it generates the
timestamp for the v4 signing so this pins to the revision with the bug
fix.