Declare Record names for indexing
Adds consistent elasticsearch `_id` so reindexing will overwrite a record and not add a new one.

This also adds support for the naming we expect on the API side. This isn't strictly necessary, but I think it will be helpful for debugging purposes in the future if the elasticsearch has the same fieldnames as the API responses. As such, this is part of DIP-164.

This removes `Record.URL` as we will be handling that as part of Record.Links.

This does not create index mappings which is the crux of what DIP-167 needs to do so this does not close the ticket even if at first look they appear related.