Add Dockerfile
This adds a Dockerfile to build the mario app. In order to keep the
build context clean, this also moves the rules file and language file to
a new config directory.

The container provides an unmodified entrypoint to the mario command so
it will work exactly the same as if it were run locally. The one thing
to keep in mind is that passing in non-default config files (rules,
languages, etc.) at run time will have to use volumes.