Add delete index functionality
This adds a delete command to the cli that works just like the create
counterpart. We should *really* consider changing this soon because
there are no guard rails in place for this operation, and by default it
will operate on timdex.

Adding this now because we currently have no way to delete an index on
AWS.