Dates are the best
MARC records use non-datelike objects for dates. We tried to treat them like dates in ElasticSearch and it blew up.

This punts and treats them as text.

Future work might try to add a separate facetable field that takes the known aspects of date oddities in MARC into account, but for now we just need to get these records into the index.