ArchivesSpace EAD processing
- processes XML as a stream and maps EAD to Record.

Stream XML parsing based on this:
https://blog.singleton.io/posts/2012-06-19-parsing-huge-xml-files-with-go/

- adds Citation per our original spec.

- Adds a new `silent` consumer that is useful in debugging as you
can dump data in part of the processing pipeline without it being lost
in the output from a normal consumer.

- Currently only keeps http prefixed Link URLs
There is a lot of data cleanup necessary and this will allow us to
proceed with links we are reasonably confident are links.

Future related work:
https://github.com/MITLibraries/mario/issues/100

- Holdings needs future work:
https://github.com/MITLibraries/mario/issues/101