Add new CLI and new bucket ADRs
This proposes to replace a couple ADRs. The first defines the structure
of the CLI, but more in terms of guidelines than explicitly listing out
commands. As the CLI will likely continue to evolve, this ADR should
remain relevant with those changes.

The second ADR redefines our usage of S3 buckets, shifting to a single
bucket for all source data.