Add a new Elasticsearch client
This adds a new client that wraps the underlying Elasticsearch client.
The goal here is to address two existing problems. The first is that a
lot of business logic has found its way into main.go and it needs to not
live there. This moves most of the logic around indexes into this
client. The second problem is that the direct usage of the underlying
client makes it difficult to unit test parts of the application. By
encapsulating that in a trimmed down client we can more easily mock an
Elasticsearch instance.

The client also more explicitly enforces business rules for our indexing
strategy. Retrieving the current index for a particular source will
return an error if more than one index is found. Promoting a new index
is now an atomic action. These steps should prevent us from reaching a
state where the primary alias points to more than index from one source.

It's worth noting that this simply adds the client. The rest of the app
will need to be updated to use it.