Refactor project layout and add ingester
There are two main changes in this commit. The first is to refactor the
layout of the application to the commonly used pkg structure. The second
moves most of the business logic that had accumulated in main.go into a
new ingester. In doing so this also switches to the new Elasticsearch
client.