Use pkger for configuration files
This switches to using pkger for handling the non-go data files that are
currently stored in the config/ directory. There are three important
implications of this change:

1. When opening a file you should use pkger.Open instead of os.Open and
use an absolute path relative to the project's root directory, for
example: pkger.Open("/config/somefile.txt"). The function signature is
the same as for os.Open.
2. The current mario command allows for passing a rules file as an
argument. For now, I have changed the default value to a path that works
for pkger, but this feature should probably either be removed entirely
or we will need to find some other solution.
3. The static files are "compiled" using the pkger command into a file
named pkged.go in project root. During local development I think pkger
should pick up changes to the file without needing to run pkger (still a
little unclear on this). It definitely needs to be run before each build
during CI/CD though. The safest route will be to always use make to
build the project.

I'm undecided on whether or not pkged.go should be kept under version
control. I have left it out for now.