Fix broken docker image
There were a few things wrong with the docker image as a result of
switching to pkger. The first is that it apparently needs underlying go
tooling on the system, so the base build for the production container
was switched to the golang alpine image. The second is mostly that I
still don't really understand how pkger works. I think the problem is
probably due to the interactions between pkger and go modules. More work
to understand what's going on here is needed, but for now copying the
config files to the final layer is a quick fix.