Switch to Github Actions
This removes Travis and adds Github Actions. The test action will run on
every push to any branch except master. It runs the tests and makes sure
the docker image will build. The stage action will run on any push to
master, i.e. merging a PR. It firsts runs the tests, then builds the
docker image and pushes it to the staging repo in AWS.