Replace logging that was removed
Adds back the logging that was removed during
the application refactor last summer. This does
not add new logging or address any other issues
around the refactor.

Why these changes are being introduced:
The refactor removed a few lines of logging that are valuable for
checking whether Aleph updates are indexing properly. This just returns
them to the codebase.

How this addresses that need:
* Adds back a few lines of logging that were removed from the original
  main.go (they are now in main.go and ingester.go).

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/DISCO-204