Remove default prefix from Mario commands
Why these changes are being introduced:
With a default prefix of "aleph", it was easy to accidentally promote
one index and remove an index of a different type. This removes the
default value and sets prefix to be a required option for both the
ingest and promote commands in Mario.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/DISCO-198