Consolidates source/type and prefix in cli
Why these changes are being introduced:

This addresses two issues: first is that source (aleph, dspace, etc.)
and type (marc, json, etc.) were conflated in the cli "type" option.
Once that was addressed and source is used consistently throughout the
app, source and prefix were redundant options as they are both always
the same in practice.

How this addresses that need:
* Renames "type" to "source" in the cli and updates the source options
  throughout the application
* Removes the "prefix" option in the cli and uses the source instead

Side effects of this change:

All external services using the cli (including DSpace and ASpace
airflow tasks and the Aleph lambda) will need to be updated.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/DISCO-260