Clean up CLI
Why these changes are being introduced:
The CLI has been a bit messy and confusing for some time as
documented in ADR 14. This finishes the work to bring the
CLI in line with that ADR.

How this addresses that need:
* Removes index from global options
* Organizes all commands in two groups: ES actions and index actions
* Adds "required" to all required options
* Reorganizes ingest options to have index first (to match the other
  index commands) followed by required options, then optional ones
* Removes RulesFile option from ingest command, also removes it from
  the ingester configuration, and instead hard codes the rules file
  directly in the MARC generator (to match other MARC-specific
  config file handling)
* Removes prefix option from the promote command and determines the
  prefix from the index name within the es Promote function
* Fixes a bug in the es Promote function that would demote an index
  if it was already in the timdex-prod alias and it was passed to the
  promote command
* Updates README command examples to match the new cli

Side effects of this change:
As noted in the ticket, the Aleph and ASpace automated ingest
infrastructure needs to be redeployed with these updated commands.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/DISCO-99