Enforce index naming pattern
Why these changes are being introduced:
It was previously possible to provide an arbitrary index name to the
ingest command, which would create a new index with that name. This
change checks the provided index name against the existing indexes
and returns an error if there is no current index with the provided
name. This effectively prevents a user from creating an index with a
non-standardized name by forcing all new indexes to have names
generated by the system.

Side effects of this change:
No side effects, but a note that this does not inherently resolve the
current staging/production issue where we have at least one prod index
that does not follow our standard naming pattern. That issue will need
to be resolved manually, and then should not be a problem going forward.