Change all references to aleph
Why these changes are being introduced:
Our ingest command and some of the documentation referenced Aleph as
the source system for MARC records. Now that we're migrating to Alma,
we need to change all of those references to use the current system
name.

How this addresses that need:
* Changes the CLI source option for MARC records from aleph to alma
* Updates references to Aleph in documentation and charts

Side effects of this change:
* Mario-powerup command needs to be changed to match the new CLI source
  option (already handled in a PR to that repo).
* Note that ADRs referencing Aleph remain, as those are accurate records
  of our decisions at the time.