Fix bug in Dockerfile
Why these changes are being introduced:
For reasons currently unknown, updates in a previous commit caused
problems with the Docker build that prevented running the Mario command
in the container. CI build on merge was also broken due to previous
renaming of the main branch.

How this addresses that need:
* Dockerfile updated to result in a successful build that will run a
  Mario command
* Github Actions config files now reference the main branch instead of
  master