Revert Go version to 1.15
We ran into issues with the Docker build using Go version 1.16-alpine.
It's still unclear exactly what cause the issues, but the built image
successfully runs Mario commands using 1.15 so we're reverting to that
for now. A new ticket will address upgrading to Go 1.16.

Note: reverting all the way back to 1.13 was not an option, as the
1.13-alpine Docker image is no longer supported. 1.15 is the latest
working version.