Add multistage build back to 116
# Subject
Switch to Multistage build of docker container

# Why these changes are being introduced:
These changes are based on this stackoverflow -
https://stackoverflow.com/questions/66894200/go-go-mod-file-not-found-in-current-directory-or-any-parent-directory-see-go

Which indicaties that mod is more necessary than before in the go configuration.
By being verbose in using the workdir and including the go.mod it seems to work again.

# How this addresses that need:
It makes 116 work with multistage builds again.

# Side effects of this change:
# None

# Relevant ticket(s):
# * https://mitlibraries.atlassian.net/browse/DISCO-291