Update for OpenSearch
Why these changes are being introduced:
We've decided to use OpenSearch instead of Elasticsearch moving
forward. The repo had a previous ADR with the decision to use
Elasticsearch, along with many references in the code comments and
Readme to Elasticsearch. These all need updating to reflect the
decision.

Note: the code itself works as is against OpenSearch (after the last
commit with updates to work against Elasticsearch v7), so the changes
in this commit are purely around documentation.

How this addresses that need:
* Adds ADR 16 to supercede ADR 2, replace Elasticsearch with OpenSearch
* Replaces all code comments and charts referencing Elasticsearch with
  OpenSearch
* Updates README to specify OpenSearch Docker image instead of
  Elasticsearch for local development

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-25