Update Github actions to deploy from 2.x branch
Why these changes are being introduced:
While we work on breaking changes that will result in version 3.0 of
Mario, we want to retain a long-running branch for version 2.x that can
be deployed to production until 3.0 is ready to go live.

How this addresses that need:
* Updates the staging Github actions workflow to deploy the 2.x branch
* Updates the test Github actions workflow to ignore the 2.x branch in
  addition to main
* Updates both workflows to use our current Go version, 1.16

Side effects of this change:
All future work intended for immediate production release should be
PR'd into the 2.x branch. Work to support new features can be PR'd into
the main branch.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-39