Migrate ingest to new data model
This migrates the JSON generator and other ingest-
related functions to match the new TIMDEX data
model. Non-JSON generators are not being migrated.

Why these changes are being introduced:
As we update the TIMDEX ETL process, we are splitting up the transform
and load steps that have previously been coupled in this codebase. This
commit updates the ingest code to *only* accept ingest data in the
TIMDEX JSON format. Data is loaded directly into OpenSearch as TIMDEX
JSON, and is not transformed here. New transform processes are being
migrated to other codebase(s).

How this addresses that need:
* Updates the OpenSearch mappings document
  (config/es_record_mappings.json) and record model
  (pkg/record/record.go) to match the new TIMDEX data model
* Adds a new fixture of TIMDEX JSON sample records for testing and
  updates the jsonrecord_test to match
* Updates the elastic client to use the new TimdexRecordId field as the
  OpenSearch document ID for each indexed record
* Removes all of the non-JSON source generators and associated test
  files

Other changes:
* Update Makefile to use new ECR repository name

Side effects of this change:
Ingest from any of the previous sources (Alma, ArchivesSpace, DSpace)
is no longer possible.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-50